Add option to ignore `incrementVersion` for snapshot versions
More of a hack than anything - I want `3.0.0-SNAPSHOT` to be published right now not `3.0.1-SNAPSHOT`

The action needs a recode and preferably with code-coverage. It's an internal Action so we should
just make assumtions about behaviour instead of trying to make it generic.
Not being able to properly 'script' when defining GitHub Workflows makes for weird hacks
when values provided to actions are dependent on other variables etc.